make note and text moveable again / moved policy to policy folder
