+++ plugins/org.yakindu.sct.ui.editor/src/org/yakindu/sct/ui/editor/editparts/SynchronizationEditPart.java
@@ -21,6 +21,7 @@ import org.yakindu.base.gmf.runtime.editpolicies.BarResizeEditPolicy;
 import org.yakindu.sct.ui.editor.editor.figures.SynchronizationFigure;
 import org.yakindu.sct.ui.editor.policies.EnlargeContainerEditPolicy;
 import org.yakindu.sct.ui.editor.policies.FeedbackGraphicalNodeEditPolicy;
+import org.yakindu.sct.ui.editor.policies.FixedBendpointEditPolicy;
 
 /**
  *
