+++ plugins/org.yakindu.sct.ui.editor/src/org/yakindu/sct/ui/editor/policies/EnlargeContainerEditPolicy.java
@@ -33,7 +33,6 @@ import org.eclipse.gmf.runtime.diagram.ui.editparts.IGraphicalEditPart;
 import org.eclipse.gmf.runtime.diagram.ui.l10n.DiagramUIMessages;
 import org.eclipse.gmf.runtime.emf.core.util.EObjectAdapter;
 import org.yakindu.base.gmf.runtime.editpolicies.SetPreferredSizeRequest;
-import org.yakindu.sct.ui.editor.editparts.FixedBendpointEditPolicy;
 import org.yakindu.sct.ui.editor.editparts.RegionEditPart;
 import org.yakindu.sct.ui.editor.editparts.StateEditPart;
